fixed build failure
