+++ src/app/api/signin-with-otp/route.ts
@@ -5,15 +5,7 @@ import { serverContainer } from '@/services/server/container';
 import { SERVER_SERVICE_KEYS } from '@/services/server/keys';
 import { ServerError } from '@/errors/server-error';
 import { ZodError } from 'zod';
-
-export const rateLimiter = serverContainer.get(
-  SERVER_SERVICE_KEYS.createRateLimiter,
-)({
-  route: '/signin-with-otp',
-  allowedRequests: 10,
-  duration: 1,
-  durationUnit: 'd',
-});
+import { rateLimiter } from './rate-limiter';
 
 export async function POST(request: NextRequest) {
   /* istanbul ignore next */
